Drop the ability to chain assertions.
I've unfortunately found it quite difficult to leave assertion chaining in whilst also providing for modifiers for existing assertions (such as exact() for contains(...)). I'd need to implement Into<Spec> to allow for any extended Spec struct to also be compatible with the existing assertions, but it appears that inference for generic traits containing generic structs does not support that use case.

So, given the choice of the two features, modifiers are likely the better feature to have.