Remove self return from is_equal_to and is_not_equal_to
Looks like they were missed when I was removing all of the other assertion returns earlier. Unfortunately this is a breaking change, but hopefully no-one is actually relying upon this functionality as it isn't documented, and the existing documentation is already reasonably specific about when you can chain assertions together.

Either way, the change has to be made at some point.